requirements.txt
ydb==3.8.0
pandas==2.2.0
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.1
regex==2023.12.25


